Don't ignore arguments to `hole` rule
Previously, Carcara didn't parse arguments to `hole` steps, to allow
for arbitrary arguments to be passed. However, if a proof is using
sharing and a term is named in the arguments of a `hole` step, Carcara
would return an error when that name was used later, since the name was
never defined.

Now, Carcara will fully parse all arguments to `hole` steps, and the
proof producers must ensure that these arguments are valid terms.